+++ nessi/signal/tapering.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 # -------------------------------------------------------------------
-# Filename: windowing.py
+# Filename: tapering.py
 #   Author: Damien Pageot
 #    Email: nessi.develop@protonmail.com
 #
@@ -25,9 +25,59 @@ def _linear(n, ntap1, ntap2):
 
     # Create the taper function
     for i in range(0, ntap1):
-        ftap[i] = float(i)/float(ntap1-1)
-    for i in range(n-ntap2+1, n):
-        ftap[i] = float(n-i-1)/float(ntap2-1)
+        if i == 0:
+            ftap[i] = 0.
+        else:
+            ftap[i] = float(i)/float(ntap1)
+    for i in range(0, ntap2):
+        if i == 0:
+            ftap[n-i-1] = 0.
+        else:
+            ftap[n-i-1] = float(i)/float(ntap2)
+
+    return ftap
+
+def _sine(n, ntap1, ntap2):
+    """
+    Sine taper type.
+    """
+    # Initialize taper function
+    ftap = np.zeros(n, dtype=np.float32)
+    ftap[:] = 1
+
+    # Create the taper function
+    for i in range(0, ntap1):
+        if i == 0:
+            ftap[i] = 0.
+        else:
+            ftap[i] = np.sin(np.pi*float(i)/float(ntap1)/2.)
+    for i in range(0, ntap2):
+        if i == 0:
+            ftap[n-i-1] = 0.
+        else:
+            ftap[n-i-1] = np.sin(np.pi*float(i)/float(ntap2)/2.)
+
+    return ftap
+
+def _cosine(n, ntap1, ntap2):
+    """
+    Cosine taper type.
+    """
+    # Initialize taper function
+    ftap = np.zeros(n, dtype=np.float32)
+    ftap[:] = 1
+
+    # Create the taper function
+    for i in range(0, ntap1):
+        if i == 0:
+            ftap[i] = 0.
+        else:
+            ftap[i] = 0.5*(1.0-np.cos(np.pi*float(i)/float(ntap1)))
+    for i in range(0, ntap2):
+        if i == 0:
+            ftap[n-i-1] = 0.
+        else:
+            ftap[n-i-1] = 0.5*(1.0-np.cos(np.pi*float(i)/float(ntap2)))
 
     return ftap
 
@@ -48,7 +98,12 @@ def taper1d(dobs, ntap1, ntap2, min, type, axis=0):
             n = np.size(dobs, axis=1)
 
     # Calculate the taper function
-    ftap = _linear(n, ntap1, ntap2)
+    if type == 'linear':
+        ftap = _linear(n, ntap1, ntap2)
+    if type == 'sine':
+        ftap = _sine(n, ntap1, ntap2)
+    if type == 'cosine':
+        ftap = _cosine(n, ntap1, ntap2)
 
     # Apply the taper function
     if n == 1: